// frontend/lib/api/generated/model/inspeccionDetalleIn.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * SIGA-UNMSM API
 * API del Sistema Integral de Gestión de Almacén y Abastecimiento Alimentario — UNMSM. Módulos activos: Autenticación, Usuarios, Almacenes, Catálogo Nutricional.
 * OpenAPI spec version: 0.1.0
 */

export interface InspeccionDetalleIn {
  guia_remision_detalle_id: number;
  /** @minimum 0 */
  cantidad_conforme: number;
  /** @minimum 0 */
  cantidad_observada: number;
}